requirements.txt
fastmcp>=2.10.0
openai==1.55.3
pydantic==2.11.7
pillow==10.4.0
python-dotenv==1.1.0
uvicorn==0.31.1
aiofiles==24.1.0
httpx==0.28.1
requests==2.31.0
